Change map behavior to respond to url hash
This commit changes a major aspect of the map behavior, enabling it to
respond to location hashes.

The GeoMap class now has a method called zoomToHash(), which reads the
window.location.hash property and then cross-references that to the
geoJson data, zooming the map and adding a marker if a location is
found. A new addOpenPopup() method has also been added to facilitate
this. There is a global RegionMap object exposed as a reference to
faciliate this feature (RegionMap.map is the leaflet instance itself).
On pages that have a map element, a window.onhashchange listener calls
this function when the hash updates.

This should account for both changing hashes on the same page as a map,
as well as following links to a different page to a map location.

In the text files for chapters, locations should be referenced as
follows:
`<%= link_to "the Bay of Naples", "/catalogue/italy.html#loc_21352" %>`

The map page should be specified manually (i.e. whether the location
should appear on the Italy map, Syria map, etc.).