Add location_link helper method
This commit enables the location_link helper method and removes a
strange CSS bug where superscript elements were getting the paragraph
indent style for some reason.

To use this method in the text, follow this format:

<%= location_link("Kisamos", "/catalogue/italy.html#loc_23373") %>

Where the first argument is the name which will appear, and the second
argument is the complete path of the HTML page (not the final URL, so
use .html ending here) plus the anchor link of loc_MAPID.

An example appears in the Kisamos entry on Cat 1.