Add section-specific print maps
This commit adds the "slugify" gem as a dependency.

The print version of the book now contains section-specific maps. The
partial which renders these knows which map to use by stashing the
region as a data-map attribute with value ==
current_page.data.title.slugify.

This means the print stylesheet can look for a region-map element where:
&[data-map="italy"], etc.