Use asset paths for apple-touch-icons
Middleman's templates support a `favicon_tag` helper method which is
asset-pipeline-aware. This means that the correct URL for the icons will appear
on every page of the site, not just the home page.

Here's how to use this method in a HAML template:

/ Apple home screen icons
= favicon_tag 'apple-touch-icon-57x57.png', rel: 'apple-touch-icon', type: 'image/png', sizes: '57x57'
= favicon_tag 'apple-touch-icon-72x27.png', rel: 'apple-touch-icon', type: 'image/png', sizes: '57x57'
= favicon_tag 'apple-touch-icon-114x114.png', rel: 'apple-touch-icon', type: 'image/png', sizes: '114x114'
= favicon_tag 'apple-touch-icon-144x144.png', rel: 'apple-touch-icon', type: 'image/png', sizes: '144x144'

The icons need to live in /assets/images under the same names used as the first
argument in order to be found.